stock_control_backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.8.0
Django==5.2
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
